Started to do a little too much in one go = activated all comopiler warnings + started to adress them - but actual contribution is that RootState now handles a code point buffer + update and view dispatch to state (not used for now)

+++ src/csv/statement_table_meta.cpp
@@ -6,8 +6,8 @@
 namespace SKV {
   std::optional<OrgNo> to_org_no(std::string_view sv) {
     auto len = sv.size();
-    int digits_count{};
-    int hyphen_count{};
+    uint8_t digits_count{};
+    uint8_t hyphen_count{};
     for (size_t i=0;i<sv.size();++i) {
       auto ch = sv[i];
       if (std::isdigit(ch)) ++digits_count;
